[unfinished] Inline `ExpressionConvertorVisitor.java`; Refactor grammar
